+++ packages/core-js/modules/web.url-search-params.js
@@ -117,7 +117,7 @@ var URLSearchParamsIterator = createIteratorConstructor(function Iterator(params
   if (!step.done) {
     step.value = kind === 'keys' ? entry.key : kind === 'values' ? entry.value : [entry.key, entry.value];
   } return step;
-});
+}, true);
 
 var URLSearchParamsState = function (init) {
   this.entries = [];