+++ tests/tests/web.url-search-params.js
@@ -1,6 +1,8 @@
 import { DESCRIPTORS, NODE } from '../helpers/constants';
 import { createIterable } from '../helpers/helpers';
 
+const { getPrototypeOf, getOwnPropertyDescriptor } = Object;
+
 QUnit.test('URLSearchParams', assert => {
   assert.isFunction(URLSearchParams);
   assert.arity(URLSearchParams, 0);
@@ -703,6 +705,8 @@ QUnit.test('URLSearchParams#entries', assert => {
     result += key + value;
   }
   assert.same(result, 'a1c3');
+
+  assert.true(getOwnPropertyDescriptor(getPrototypeOf(new URLSearchParams().entries()), 'next').enumerable, 'enumerable .next');
 });
 
 QUnit.test('URLSearchParams#keys', assert => {
@@ -746,6 +750,8 @@ QUnit.test('URLSearchParams#keys', assert => {
     result += key;
   }
   assert.same(result, 'ac');
+
+  assert.true(getOwnPropertyDescriptor(getPrototypeOf(new URLSearchParams().keys()), 'next').enumerable, 'enumerable .next');
 });
 
 QUnit.test('URLSearchParams#values', assert => {
@@ -789,6 +795,8 @@ QUnit.test('URLSearchParams#values', assert => {
     result += key;
   }
   assert.same(result, '13');
+
+  assert.true(getOwnPropertyDescriptor(getPrototypeOf(new URLSearchParams().values()), 'next').enumerable, 'enumerable .next');
 });
 
 QUnit.test('URLSearchParams#@@iterator', assert => {
@@ -838,6 +846,8 @@ QUnit.test('URLSearchParams#@@iterator', assert => {
     result += key + value;
   }
   assert.same(result, 'a1c3');
+
+  assert.true(getOwnPropertyDescriptor(getPrototypeOf(new URLSearchParams()[Symbol.iterator]()), 'next').enumerable, 'enumerable .next');
 });
 
 QUnit.test('URLSearchParams#@@toStringTag', assert => {